Support match expression v2 rfc
This uses a version of the match expression v2 RFC that was
manually rebased against the latest 8.0-dev commits:
https://github.com/TysonAndre/php-src/pull/new/match_expression_v2

The naming is based on `switch` statements.
Note that even if statement blocks were supported after a future RFC
those would probably be considered block **expressions**,
hence the `expr` name for `AST_MATCH_ARM`.